# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPCMS \u2014 Course Project Configuration and Management System",
    "description": "Internal University Course Project SCM & Configuration Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_v1_projects__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/proposals": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "summary": "List Proposals",
        "operationId": "list_proposals_api_v1_teams__team_id__proposals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response List Proposals Api V1 Teams  Team Id  Proposals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Submit Proposal",
        "operationId": "submit_proposal_api_v1_teams__team_id__proposals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{id}/decision": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Decide Proposal",
        "operationId": "decide_proposal_api_v1_proposals__id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutcomeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "position",
          "text",
          "created_at"
        ],
        "title": "OutcomeOut"
      },
...
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "faculty_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty User Id"
          },
          "faculty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semester": {
            "type": "string",
            "title": "Semester"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "type": "integer",
            "title": "Min Team Size"
          },
          "max_team_size": {
            "type": "integer",
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "type": "string",
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "type": "integer",
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "type": "integer",
            "title": "Max File Mb"
          },
          "late_policy": {
            "type": "string",
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "type": "number",
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "type": "boolean",
            "title": "Versioning Required"
          },
          "git_required": {
            "type": "boolean",
            "title": "Git Required"
          },
          "baseline_frequency": {
            "type": "string",
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "type": "boolean",
            "title": "Change Request Required"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required"
          },
          "branching_policy": {
            "type": "string",
            "title": "Branching Policy"
          },
          "total_marks": {
            "type": "integer",
            "title": "Total Marks"
          },
          "grade_bands": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Grade Bands"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/ThemeOut"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeOut"
            },
            "type": "array",
            "title": "Outcomes",
            "default": []
          },
          "ci_templates": {
            "items": {
              "$ref": "#/components/schemas/CITemplateOut"
            },
            "type": "array",
            "title": "Ci Templates",
            "default": []
          },
          "deadlines": {
            "items": {
              "$ref": "#/components/schemas/DeadlineOut"
            },
            "type": "array",
            "title": "Deadlines",
            "default": []
          },
          "evaluation_criteria": {
            "items": {
              "$ref": "#/components/schemas/CriteriaOut"
            },
            "type": "array",
            "title": "Evaluation Criteria",
            "default": []
          },
          "user_team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Id"
          },
          "user_team_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Number"
          },
          "user_team_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "faculty_user_id",
          "name",
          "semester",
          "academic_year",
          "min_team_size",
          "max_team_size",
          "team_formation_mode",
          "max_projects_per_student",
          "allowed_file_exts",
          "max_file_mb",
          "late_policy",
          "late_penalty_percent",
          "versioning_required",
          "git_required",
          "baseline_frequency",
          "change_request_required",
          "approval_required",
          "branching_policy",
          "total_marks",
          "grade_bands",
          "status",
          "created_at"
        ],
        "title": "ProjectOut"
      },
...
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "semester": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "academic_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Team Size"
          },
          "max_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max File Mb"
          },
          "late_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versioning Required"
          },
          "git_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Required"
          },
          "baseline_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Request Required"
          },
          "approval_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Required"
          },
          "branching_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branching Policy"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "grade_bands": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Bands"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
      "ProposalCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 3,
            "title": "Title"
          },
          "abstract": {
            "type": "string",
            "minLength": 10,
            "title": "Abstract"
          },
          "problem_statement": {
            "type": "string",
            "minLength": 10,
            "title": "Problem Statement"
          },
          "objectives": {
            "type": "string",
            "minLength": 10,
            "title": "Objectives"
          },
          "theme_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Id"
          },
          "expected_outcome": {
            "type": "string",
            "minLength": 10,
            "title": "Expected Outcome"
          }
        },
        "type": "object",
        "required": [
          "title",
          "abstract",
          "problem_statement",
          "objectives",
          "expected_outcome"
        ],
        "title": "ProposalCreate"
      },
      "ProposalDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(approved|rejected|revision_required)$",
            "title": "Decision"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ProposalDecisionRequest"
      },
      "ProposalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "revision_no": {
            "type": "integer",
            "title": "Revision No"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "abstract": {
            "type": "string",
            "title": "Abstract"
          },
          "problem_statement": {
            "type": "string",
            "title": "Problem Statement"
          },
          "objectives": {
            "type": "string",
            "title": "Objectives"
          },
          "theme_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Id"
          },
          "theme_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Name"
          },
          "expected_outcome": {
            "type": "string",
            "title": "Expected Outcome"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "faculty_feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Feedback"
          },
          "submitted_by": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By"
          },
          "submitted_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Name"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By Name"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "revision_no",
          "title",
          "abstract",
          "problem_statement",
          "objectives",
          "expected_outcome",
          "status",
          "submitted_by",
          "created_at"
        ],
        "title": "ProposalOut"
      },
...
      "ThemeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "created_at"
        ],
        "title": "ThemeOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}